Rename MenuGeneration classes
