+++ python/tk_3dsmax/__init__.py
@@ -8,6 +8,6 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
+from .menu_generation_menuman import MenuGenerator_menuMan
 from .menu_generation import MenuGenerator
-from .menu_generation_2025 import MenuGenerator2025
 from .maxscript import MaxScript
